Add basic support for versioning
Previous versions of SADFace are considered to be version 0.1 and
the current version as of this commit is version 0.2 which will
begin the process of introducing minor breaking changes. In the
long term, versioning will move to a semantic versioning system
e.g. MAJOR.MINOR.BUGFIX numbering but until version 1.0 we'll use
even version numbers to indicate changes that break backwards
compatibility